cyclesafe-backend/cyclesafe_backend/users/urls.py
from django.urls import path
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView
from .views import RegisterView, PasswordResetView, NotificationSettingsView

urlpatterns = [
    #JWT Authentication Endpoints
    path('token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),

    #  User Management
    path('register/', RegisterView.as_view(), name='register'),
    path('password-reset/', PasswordResetView.as_view(), name='password_reset'),

    # Notification Settings (Phone + SMS)
    path('notifications/', NotificationSettingsView.as_view(), name='notification_settings'),
]
